+++ conjur/logout/__init__.py
@@ -1,0 +1,8 @@
+"""
+Logout module
+
+This metafile includes all the functionality that will be exposed
+when you install this module
+"""
+from .logout_logic import LogoutLogic
+from .logout_controller import LogoutController