+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 
 ### Added
+- Method `login\logout` to login and out of the CLI
+  [cyberark/conjur-api-python3#128](https://github.com/cyberark/conjur-api-python3/issues/128)
 - Method `init` to initialize the CLI with the Conjur server is now available in both CLI and API
   [cyberark/conjur-api-python3#89](https://github.com/cyberark/conjur-api-python3/issues/89)